requirements.txt
# Pinned to versions confirmed working as of 2026-05-10.
# Install with: pip install -r requirements.txt
#
# Why pin versions: unpinned requirements mean any future major release
# of any package could break the app silently on a fresh install.
# For a single-developer project, pinned requirements.txt achieves the
# same reproducibility goal as pyproject.toml with a lockfile, with
# less setup overhead.

# --- Core app ---
streamlit==1.56.0
python-dotenv==1.2.2
requests==2.33.1

# --- LangChain / LangGraph stack ---
langchain==1.2.17
langchain-openai==1.2.1
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-core==1.3.3
langchain-text-splitters==1.1.2
langchain-huggingface==1.2.2
langgraph==1.1.10

# --- Vector store and embeddings ---
chromadb==1.5.9
sentence-transformers==5.4.1
protobuf==5.29.4

# --- Cost tracking ---
tiktoken==0.12.0

# --- Evaluation (RAGAS + source-hit) ---
ragas==0.4.3
datasets==4.8.5
pandas==2.3.3

# --- Testing ---
pytest==9.0.3
